// crates/desktop/src/generated/dto/ProviderTransferItemDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ProviderTransferStatusDto } from "./ProviderTransferStatusDto";

export type ProviderTransferItemDto = {
	name: string;
	status: ProviderTransferStatusDto;
	reason: string | null;
};